% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely.R
\name{safely}
\alias{safely}
\title{Make a safe version of a function}
\usage{
safely(.f)
}
\arguments{
\item{.f}{a function to wrap}
}
\description{
`safely()` wraps its input function, creating a new function that never
throws an error. Instead, it returns an object of class `safely` which
is a list with two components, `result` and `error`. One component is
always null
}
\examples{
safe_log <- safely(log)
safe_log(10)
safe_log("10")
}
